Remove deprecated util functions
https://github.com/crossplaneio/crossplane-runtime/issues/1

Each of these functions were called by only a single package, so I've raised
PRs to move them to the packages that call them. All relevant PRs are cross
linked from the above issue
