--- pkg/util/retry.go
@@ -1,32 +1,0 @@
-/*
-Copyright 2019 The Crossplane Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package util
-
-import (
-	"time"
-
-	"k8s.io/apimachinery/pkg/util/wait"
-)
-
-// DefaultUpdateRetry is the recommended retry for a failure to update a resource or ensuring that the changes
-// from an update have been committed and have become consistent.
-var DefaultUpdateRetry = wait.Backoff{
-	Steps:    10,
-	Duration: 500 * time.Millisecond,
-	Factor:   1.0,
-	Jitter:   0.1,
-}
